feat(schema): change id and add createdAt for historicalBalance entity
