+++ schema.graphql
@@ -20,13 +20,15 @@ type LBPPool implements Pool @entity {
     historicalBalances: [HistoricalBalance!]! @derivedFrom(field: "pool")
 }
 
+# TODO generalize LBPPool reference to Pool; consider using union for XYKPool | LBPPool
 type HistoricalBalance @entity {
-  "paraChainBlockHeight"
-  id: ID!
-  pool: LBPPool!
-  assetABalance: BigInt!
-  assetBBalance: BigInt!
-  relayChainBlockHeight: BigInt!
+    "PoolId-paraChainBlockHeight"
+    id: ID!
+    pool: LBPPool!
+    assetABalance: BigInt!
+    assetBBalance: BigInt!
+    relayChainBlockHeight: BigInt!
+    createdAt: DateTime!
 }
 
 type Chronicle @entity {
